feat: Add search functionality to SearchFunctionalityController
Add the search functionality to the SearchFunctionalityController. This includes
the addition of the `@GetMapping` annotation with the `params` parameter set to
"search". The method `searchProducts` is implemented to retrieve products from
the product repository based on the search query. The search is performed by
calling the `findByNameContainingOrDescriptionContaining` method of the
ProductRepository with the search query. The retrieved products are then mapped
to the Product class using the ModelMapper. Finally, the products are returned
in the ResponseEntity.

Also, update the getAllProducts method to retrieve all products from the
product repository and map them to the Product class using the ModelMapper.

Repository: C:/Users/olive/ws/atmosphere-service-search-and-filtering
Files:
- src/main/java/de/ithoc/atmosphere/service/searchandfiltering/api/SearchFunctionalityController.java
- src/main/java/de/ith