feat: Add searchProductsByPriceRange endpoint
Add a new GET endpoint to search products by price range. The endpoint
accepts "fromPrice" and "toPrice" as query parameters. The endpoint
fetches products from the repository that have a price between the
specified range. The fetched product entities are then mapped to
Product objects using modelMapper and returned as a list of products
in the response body.