feat: Add RequestHeaderInterceptor class
Add the RequestHeaderInterceptor class to the config package. This class is responsible for intercepting client HTTP requests and adding a custom header to the request.

The RequestHeaderInterceptor class implements the ClientHttpRequestInterceptor interface and has two constructor parameters - headerName and headerValue. These parameters define the name and value of the custom header that will be added to the request.

The intercept() method overrides the method from the ClientHttpRequestInterceptor interface. It receives the request, body, and execution objects as parameters. Inside the method, the custom header is added to the request's headers using the headerName and headerValue parameters. The modified request is then passed to the execution object's execute() method, which returns the response.

This class is useful for scenarios where you need to add custom headers to client HTTP requests.